requirements.txt
azure-cognitiveservices-vision-customvision
azure-functions